requirements.txt
# General
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.2
Pillow
djangorestframework

# Third Party
python-decouple
markdown
django-filter